feat: #737 Adds resources ids extractors
